Update fcheck to check the alignment of the struct passed to getplete. The cnames struct flag variable now a Byte.

+++ z.h
@@ -109,7 +109,7 @@
 struct cnames {
 	const char *name;
 	Byte fnum;
-	short flags;
+	Byte flags;
 	const char *doc;
 };
 #define CNAMESIZE sizeof(struct cnames)
